hw0406

+++ hw01-04/taylor.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include <stdlib.h>
 #define i32 int32_t
-i32 factorio(i32 n){
+double factorio(i32 n){
     if(n==1){
         return 1;
     }else{
@@ -10,9 +10,9 @@ i32 factorio(i32 n){
 }
 double natural_log(i32 k){
     if(k<=1){
-        return 2.0;
+        return 2;
     }else{
-        return 1.0/factorio(k)+natural_log(k-1); 
+        return 1/factorio(k)+natural_log(k-1); 
     }
 }
